rename old ApplicationTypes and rename v2
